+++ packages/flow-trigger-bundle/src/types/trigger.ts
